Cleaning QMCMain hierarchy
Remove CoupledMC
Remove QMCMainState unused variables
