+++ src/QMCApp/CMakeLists.txt
@@ -18,7 +18,7 @@
 ####################################
 # create libqmc
 ####################################
-set(QMCAPPDIR QMCAppBase.cpp QMCMainState.cpp QMCMain.cpp CoupledMC.cpp)
+set(QMCAPPDIR QMCAppBase.cpp QMCMainState.cpp QMCMain.cpp)
 if(USE_OBJECT_TARGET)
   add_library(qmc OBJECT ${QMCAPPDIR})
 else()
